primitives分割 Phase 3: strings.rs を切り出す
文字列系 primitive を src/primitives/strings.rs に移動し、
resolve_str_cell を廃止して expect_str / expect_str_rc に置き換える。
